ATUALIZACAO-SNAPSHOT-VERSION-02--> IMPLEMENTADA A FUTURE DE DIMENCIONAMENTO-DE SISTEMA COMPLETO OFFGRID.

+++ src/main/java/br/com/dasare/solarOffGrid/domain/offgrid/interfac/CalculateBatteryBankInterface.java
@@ -2,7 +2,7 @@ package br.com.dasare.solarOffGrid.domain.offgrid.interfac;
 
 public interface CalculateBatteryBankInterface {
 
-    Integer BatteryBankCapacityUseful(Double totalPowerSistemPerDay, Integer autonomyDay, Integer voltagyBatteryBank);
+    Integer BatteryBankCapacityUseful(Double totalPowerSistemPerDay, Double autonomyDay, Integer voltagyBatteryBank);
     Integer batteryBankCapacityReal(Integer batteryBankCapacityUseful, Double dischargeDepthOfBatteryBank);
     Integer numberOfBatteriesInsires(Integer voltagyInInversor, Integer voltagyBattery);
     Integer numberOfBatteriesParallel(Integer batteryBankCapacityReal, Integer capacityBateryAH);
